lakegallery/requirements.txt
Django==1.11.4
django-filter==1.0.4
django-geojson==2.10.0
django-leaflet==0.22.0
djangorestframework==3.6.4
drf-extensions==0.3.1
Markdown==2.6.9
psycopg2==2.7.3.1
pytz==2017.2
six==1.10.0
